requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.9.2
django-haystack==2.4.0
django-ssl-redirect==2.1.1
django-taggit==0.17.6
idna==2.7
Markdown==2.6.2
pycparser==2.19
PyMySQL==0.9.2
pysolr==3.3.2
pytz==2018.5
requests==2.20.0
six==1.11.0
urllib3==1.24
